pyproject.toml
# pyproject.toml

[project]
name = "hrunobot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "yarl (>=1.22.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.11.0,<3.0.0)",
    "aiogram (>=3.23.0,<4.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "redis (>=7.1.0,<8.0.0)",
    "colorama (>=0.4.6,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
